requirements.txt
python-dotenv
requests
faiss-cpu
sentence-transformers
pypdf
langchain
gradio
